heis_driver v0.1.0 Driver

An elixir-driver communicating with 'SimElevatorServer' or the elevator server.

Description

You must start the driver with start_link() or start_link(ip_address, port) before any of the other functions will work. The user is responsible for not giving stupid input, or polling nonexistent buttons

Including the driver in projects

Elixir

Modify the deps-list in mix.exs, do that it includes

defp deps do
    [
      {:heis_driver, git: "https://github.com/jornbh/heis_driver.git", tag: "0.1.0"}
    ]
  end

Erlang

You need some extra plugins to include elixir-code.

  {plugins, [rebar_mix]}.
  {provider_hooks, [{post, [{compile, {mix, consolidate_protocols}}]}]}.

This makes it possible to include elixir-dependencies. Afterwards, modify the line for deps in rebar.config, so that it becomes:

  {deps, [
      {heis_driver, {git, "git://github.com/jornbh/heis_driver.git", {tag, "0.1.0"}}}
  ]}.

You might have to call the dunctions like 'Elixir.Driver':'button_pressed?'(1, hall_up) if the names do not follow the conventions for atoms in Erlang.

Further reading

GenServers are a really neat way to make servers without having to rewrite the same code all the time. It works Exactly the same in erlang as well, but it is called gen_server instead. The erlang documentation is kind of hard understand, so use the elixir-video and "Translate" it to erlang (gen_server:call(...) instead of GenServer.call(...)).

Short version is that a GenServer implements the basic parts of a server, and the code seen in this file is the "Blanks you have to fill in"

A youtube-video that explains GenServers and Supervisors

https://www.youtube.com/watch?v=3EjRvaCOl94

Credits to Jostein for implementing this

Types

button_dir()

button_dir() :: :hall_up | :hall_down | :cab

ip_address()

ip_address() :: {integer(), integer(), integer(), integer()}

motor_dir()

motor_dir() :: :up | :down | :stop

Functions

button_pressed?(floor, dir)

button_pressed?(integer(), button_dir()) :: boolean()

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

child_spec(ip, port)

child_spec(ip_address(), integer()) :: %{
  id: Driver,
  start: {Driver, :start_link, [...]}
}

poll_floor_sensor()

poll_floor_sensor() :: integer() | :between_floors

set_button_light(floor, dir, wanted_state)

set_button_light(integer(), button_dir(), :on | :off) :: :invalid_input | :ok

set_door_state(door_state)

set_floor_indicator(floor)

set_floor_indicator(integer()) :: :ok

set_motor_dir(motor_dir)

set_motor_dir(motor_dir()) :: :invalid_input | :ok

start_link()

start_link(ip, port)

start_link(ip_address(), integer()) :: {:ok, pid()}