heis_driver v0.1.0 Driver

Communicates between erlang and elixir

A driver for talking with the elevator in Erlang and Elixir

Description

You must start the driver with start_link() or start_link(ip_address, port) before any of the other functions will work. The user is responsible for not giving stupid input, or polling nonexistent buttons

API:

Functions

{:ok, driver_pid} = Driver.start_link
button_pressed?(floor, button_direction)
set_motor_dir(motor_direction)
set_button_light(floor, button_direction, on_or_off)
set_floor_indicator(floor)
set_door_state(door_state)

Data-types

door_state: (:opne/:closed) motor_direction: (:hall_up/:hall_down/:cab) floor: (0/1/.../number_of_floors -1) door_state: :open/:closed

Further reading

GenServers are a really neat way to make servers without having to rewrite the same code all the time. It works Exactly the same in erlang as well, but it is called gen_server instead. The erlang documentation is kind of hard understand, so use the elixir-video and "Translate" it to erlang (gen_server:call(...) instead of GenServer.call(...)).

Short version is that a GenServer implements the basic parts of a server, and the code seen in this file is the "Blanks you have to fill in"

A youtube-video that explains GenServers and Supervisors

https://www.youtube.com/watch?v=3EjRvaCOl94

Credits to Jostein for implementing this

Types

button_dir()

button_dir() :: :hall_up | :hall_down | :cab

ip_address()

ip_address() :: {integer(), integer(), integer(), integer()}

motor_dir()

motor_dir() :: :up | :down | :stop

state()

state() :: :on | :off

Functions

button_dir?(dir)

button_dir?(button_dir()) :: boolean()

button_pressed?(floor, dir)

button_pressed?(integer(), button_dir()) :: boolean()

check_floor()

check_floor() :: integer() | :between_floors

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

child_spec(ip, port)

child_spec(ip_address(), integer()) :: %{
  id: Driver,
  start: {Driver, :start_link, [...]}
}

set_button_light(floor, dir, wanted_state)

set_button_light(integer(), button_dir(), :on | :off) :: :invalid_input | :ok

set_door_state(door_state)

set_floor_indicator(floor)

set_floor_indicator(integer()) :: :ok

set_motor_dir(motor_dir)

set_motor_dir(motor_dir()) :: :invalid_input | :ok

start_link()

start_link(ip, port)

start_link(ip_address(), integer()) :: {:ok, pid()}